Fix in calibration: int8_t->int16_t

+++ Detectors/FIT/FT0/calibration/include/FT0Calibration/FT0CalibrationInfoObject.h
@@ -28,8 +28,8 @@ class FT0CalibrationInfoObject
   void setChannelIndex(uint8_t channel) { mChannelIndex = channel; }
   [[nodiscard]] uint8_t getChannelIndex() const { return mChannelIndex; }
 
-  void setTime(int8_t time) { mTime = time; }
-  [[nodiscard]] int8_t getTime() const { return mTime; }
+  void setTime(int16_t time) { mTime = time; }
+  [[nodiscard]] int16_t getTime() const { return mTime; }
   void setAmp(int16_t amp) { mAmp = amp; }
   [[nodiscard]] int16_t getAmp() const { return mAmp; }
